requirements.txt
fastapi
pydantic
pydantic_core
requests
pillow
torch
transformers